use static_cast; JIRA: PGAP-9540
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@102944 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/align/prosplign/Info.cpp
@@ -623,8 +623,8 @@ void CProSplignText::Output(const CSeq_align& seqalign, CScope& scope, ostream& 
 
     int npos1 = is_plus_strand?nuc_from:nuc_to;
 
-    int prot_beg_pos = protein.find_first_not_of(GAP_CHAR);
-    int prot_end_pos = protein.find_last_not_of(GAP_CHAR);
+    int prot_beg_pos = static_cast<int>(protein.find_first_not_of(GAP_CHAR));
+    int prot_end_pos = static_cast<int>(protein.find_last_not_of(GAP_CHAR));
 
     for(int i=0;i<prot_end_pos; i+=width) {
         int apos = i+width-1;
@@ -638,7 +638,7 @@ void CProSplignText::Output(const CSeq_align& seqalign, CScope& scope, ostream& 
         count(dna.begin()+i, dna.begin()+(i+width), GAP_CHAR, gaps);
         int real_bases = width-gaps;
 #else
-        int real_bases = width-count(dna.begin()+i, dna.begin()+(i+width), GAP_CHAR);
+        int real_bases = static_cast<int>(width-count(dna.begin()+i, dna.begin()+(i+width), GAP_CHAR));
 #endif
 
         int npos2 = is_plus_strand?npos1+real_bases-1:npos1-(real_bases-1);
